practice/chapter6/code/requirements.txt
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
pydantic>=2.0.0
